# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Feldera API",
...
    "license": {
      "name": "MIT OR Apache-2.0"
    },
    "version": "0.18.0"
  },
  "paths": {
...
    "/v0/programs": {
      "get": {
        "tags": [
          "Programs"
        ],
        "summary": "Fetch programs, optionally filtered by name or ID.",
        "operationId": "get_programs",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Unique program identifier.",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid",
              "nullable": true
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Unique program name.",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "with_code",
            "in": "query",
            "description": "Option to include the SQL program code or not\nin the Program objects returned by the query.\nIf false (default), the returned program object\nwill not include the code.",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of programs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgramDescr"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Specified program name or ID does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "examples": {
                  "Unknown program ID": {
                    "value": {
                      "details": {
                        "program_id": "67e55044-10b1-426f-9247-bb680e5fe0c8"
                      },
                      "error_code": "UnknownProgram",
                      "message": "Unknown program id '67e55044-10b1-426f-9247-bb680e5fe0c8'"
                    }
                  },
                  "Unknown program name": {
                    "value": {
                      "details": {
                        "name": "unknown_name"
                      },
                      "error_code": "UnknownName",
                      "message": "An entity with name unknown_name was not found"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JSON web token (JWT) or API key": []
          }
        ]
      },
      "post": {
        "tags": [
          "Programs"
        ],
        "summary": "Create a new program.",
        "operationId": "new_program",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewProgramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Program created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewProgramResponse"
                }
              }
            }
          },
          "409": {
            "description": "A program with this name already exists in the database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "details": null,
                  "error_code": "DuplicateName",
                  "message": "An entity with this name already exists"
                }
              }
            }
          }
        },
        "security": [
          {
            "JSON web token (JWT) or API key": []
          }
        ]
      }
    },
    "/v0/programs/{program_name}": {
      "get": {
        "tags": [
          "Programs"
        ],
        "summary": "Fetch a program by name.",
        "operationId": "get_program",
        "parameters": [
          {
            "name": "program_name",
            "in": "path",
            "description": "Unique program name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "with_code",
            "in": "query",
            "description": "Option to include the SQL program code or not\nin the Program objects returned by the query.\nIf false (default), the returned program object\nwill not include the code.",
            "required": false,
            "schema": {
              "type": "boolean",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Program retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgramDescr"
                }
              }
            }
          },
          "404": {
            "description": "Specified program name does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "details": {
                    "name": "unknown_name"
                  },
                  "error_code": "UnknownName",
                  "message": "An entity with name unknown_name was not found"
                }
              }
            }
          }
        },
        "security": [
          {
            "JSON web token (JWT) or API key": []
          }
        ]
      },
      "put": {
        "tags": [
          "Programs"
        ],
        "summary": "Create or replace a program.",
        "operationId": "create_or_replace_program",
        "parameters": [
          {
            "name": "program_name",
            "in": "path",
            "description": "Unique program name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrReplaceProgramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Program updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrReplaceProgramResponse"
                }
              }
            }
          },
          "201": {
            "description": "Program created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateOrReplaceProgramResponse"
                }
              }
            }
          },
          "409": {
            "description": "A program with this name already exists in the database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "details": null,
                  "error_code": "DuplicateName",
                  "message": "An entity with this name already exists"
                }
              }
            }
          }
        },
        "security": [
          {
            "JSON web token (JWT) or API key": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Programs"
        ],
        "summary": "Delete a program.",
        "description": "Deletion fails if there is at least one pipeline associated with the\nprogram.",
        "operationId": "delete_program",
        "parameters": [
          {
            "name": "program_name",
            "in": "path",
            "description": "Unique program name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Program successfully deleted"
          },
          "400": {
            "description": "Specified program is referenced by a pipeline",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "details": {
                    "program_name": "unknown_name"
                  },
                  "error_code": "ProgramInUseByPipeline",
                  "message": "Program named 'unknown_name' is in use by a pipeline"
                }
              }
            }
          },
          "404": {
            "description": "Specified program name does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "details": {
                    "name": "unknown_name"
                  },
                  "error_code": "UnknownName",
                  "message": "An entity with name unknown_name was not found"
                }
              }
            }
          }
        },
        "security": [
          {
            "JSON web token (JWT) or API key": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Programs"
        ],
        "summary": "Change one or more of a program's code, description or name.",
        "description": "If a program's code changes, any ongoing compilation gets cancelled,\nthe program status is reset to `None`, and the program version\nis incremented by 1.\n\nChanging only the program's name or description does not affect its\nversion or the compilation process.",
        "operationId": "update_program",
        "parameters": [
          {
            "name": "program_name",
            "in": "path",
            "description": "Unique program name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProgramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Program updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateProgramResponse"
                }
              }
            }
          },
          "404": {
            "description": "Specified program name does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "details": {
                    "name": "unknown_name"
                  },
                  "error_code": "UnknownName",
                  "message": "An entity with name unknown_name was not found"
                }
              }
            }
          },
          "409": {
            "description": "A program with this name already exists in the database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "details": null,
                  "error_code": "DuplicateName",
                  "message": "An entity with this name already exists"
                }
              }
            }
          }
        },
        "security": [
          {
            "JSON web token (JWT) or API key": []
          }
        ]
      }
    },
    "/v0/programs/{program_name}/compile": {
      "post": {
        "tags": [
          "Programs"
        ],
        "summary": "Deprecated. Mark a program for compilation.",
        "description": "The client can track a program's compilation status by polling the\n`/program/{program_name}` or `/programs` endpoints, and\nthen checking the `status` field of the program object.",
        "operationId": "compile_program",
        "parameters": [
          {
            "name": "program_name",
            "in": "path",
            "description": "Unique program name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompileProgramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Compilation request submitted"
          },
          "404": {
            "description": "Specified program name does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "details": {
                    "name": "unknown_name"
                  },
                  "error_code": "UnknownName",
                  "message": "An entity with name unknown_name was not found"
                }
              }
            }
          },
          "409": {
            "description": "Program version specified in the request doesn't match the latest program version in the database",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "details": {
                    "latest_version": 5
                  },
                  "error_code": "OutdatedProgramVersion",
                  "message": "Outdated program version. Latest version: '5'"
                }
              }
            }
          }
        },
        "security": [
          {
            "JSON web token (JWT) or API key": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateOrReplaceProgramRequest": {
        "type": "object",
        "description": "Request to create or replace a program.",
        "required": [
          "description",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "SQL code of the program.",
            "example": "CREATE TABLE example(name VARCHAR);"
          },
          "config": {
            "$ref": "#/components/schemas/ProgramConfig"
          },
          "description": {
            "type": "string",
            "description": "Program description.",
            "example": "Example description"
          },
          "rust_code": {
            "type": "string",
            "description": "Rust code of the program.",
            "example": "pub fn CONTAINS_NUMBER(slice: &[i32], number: i32) { for &item in slice.iter(){if item == number { return true; } } false }"
          }
        }
      },
      "CreateOrReplaceProgramResponse": {
        "type": "object",
        "description": "Response to a create or replace program request.",
        "required": [
          "program_id",
          "version"
        ],
        "properties": {
          "program_id": {
            "$ref": "#/components/schemas/ProgramId"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Information returned by REST API endpoints on error.",
        "required": [
          "message",
          "error_code",
          "details"
        ],
        "properties": {
          "details": {
            "type": "object",
            "description": "Detailed error metadata.\nThe contents of this field is determined by `error_code`."
          },
          "error_code": {
            "type": "string",
            "description": "Error code is a string that specifies this error type.",
            "example": "UnknownInputFormat"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message.",
            "example": "Unknown input format 'xml'."
          }
        }
      },
...
      "NewProgramRequest": {
        "type": "object",
        "description": "Request to create a new program.",
        "required": [
          "name",
          "description",
          "code"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "SQL code of the program.",
            "example": "CREATE TABLE example(name VARCHAR);"
          },
          "config": {
            "$ref": "#/components/schemas/ProgramConfig"
          },
          "description": {
            "type": "string",
            "description": "Program description.",
            "example": "Example description"
          },
          "name": {
            "type": "string",
            "description": "Program name.",
            "example": "example-program"
          }
        }
      },
      "NewProgramResponse": {
        "type": "object",
        "description": "Response to a new program request.",
        "required": [
          "program_id",
          "version"
        ],
        "properties": {
          "program_id": {
            "$ref": "#/components/schemas/ProgramId"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          }
        }
      },
...
      "ProgramDescr": {
        "type": "object",
        "description": "Program descriptor.",
        "required": [
          "program_id",
          "name",
          "description",
          "version",
          "status",
          "config"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "SQL code",
            "nullable": true
          },
          "config": {
            "$ref": "#/components/schemas/ProgramConfig"
          },
          "description": {
            "type": "string",
            "description": "Program description."
          },
          "name": {
            "type": "string",
            "description": "Program name (doesn't have to be unique)."
          },
          "program_id": {
            "$ref": "#/components/schemas/ProgramId"
          },
          "schema": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ProgramSchema"
              }
            ],
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/ProgramStatus"
          },
          "version": {
            "$ref": "#/components/schemas/Version"
          }
        }
      },
      "ProgramId": {
        "type": "string",
        "format": "uuid",
        "description": "Unique program id."
      },
...
    },
    "securitySchemes": {
      "JSON web token (JWT) or API key": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Use a JWT token obtained via an OAuth2/OIDC\n                               login workflow or an API key obtained via\n                               the `/v0/api-keys` endpoint."
      }
    }
  },
...
}